Add files via upload
Project description.

The data you work with will not always feel familiar —sometimes, you'll encounter data from peculiar sources, operating with peculiar measurements. Let's work with something exotic to keep you on your toes: Yandex.Realty has an archive of property ads for St. Petersburg, Russia, and the surrounding areas covering the past few years. I'll show you how to determine the market value of real estate. My task is to define parameters. This will make it possible to create an automated system that is capable of detecting anomalies and fraud.
There are two different types of data available for every apartment for sale. The first type is user input. The second type was calculated automatically based upon the map data. This includes distance from the city center, the airport, and the nearest park or body of water.